requirements.txt
Flask
flask-cors
Flask-JWT-Extended
gunicorn
pandas
scikit-learn
numpy
matplotlib
Pillow
opencv-python-headless
keras
tensorflow
fastapi
uvicorn
pybase64
typing
pydantic
pymongo
python-dotenv
bcrypt
reportlab